+++ lib/benchee.ex
@@ -59,6 +59,36 @@ for {module, moduledoc} <- [{Benchee, elixir_doc}, {:benchee, erlang_doc}] do
       end)
     end
 
+    @doc """
+    A convenience function designed for loading saved benchmarks and running formatters on them.
+
+    Basically takes the input of the map of jobs away from you and skips unnecessary steps with
+    that data missing (aka not running benchmarks, only running relative statistics).
+
+    You can use config options as normal, but some options related to benchmarking won't take
+    effect (such as `:time`).
+
+    ## Usage
+
+        Benchee.report(load: ["benchmark-*.benchee"])
+    """
+    @spec report(keyword) :: Benchee.Suite.t()
+    def report(config) do
+      unless Access.get(config, :load), do: raise_missing_load()
+
+      config
+      |> Benchee.init()
+      |> Benchee.system()
+      |> Benchee.load()
+      |> Benchee.relative_statistics()
+      |> Formatter.output()
+    end
+
+    defp raise_missing_load do
+      raise ArgumentError,
+            "You need to provide at least a :load option for report/1 to make sense"
+    end
+
     defdelegate init(), to: Benchee.Configuration
     defdelegate init(config), to: Benchee.Configuration
     defdelegate system(suite), to: Benchee.System